refactor: components 디렉터리 layout, document, common으로 분리
