--- client/src/app/page.tsx
@@ -1,3 +1,0 @@
-export default function Home() {
-  return <div className="font-bm text-4xl text-primary-600">안녕</div>;
-}
