+++ client/src/app/wiki/[title]/log/[logId]/page.tsx
@@ -1,8 +1,8 @@
 import {getSpecificDocumentLog} from '@apis/document';
-import DocumentContents from '@components/Document/DocumentContents';
-import DocumentFooter from '@components/Document/DocumentFooter';
-import DocumentHeader from '@components/Document/DocumentHeader';
-import MobileDocumentHeader from '@components/Document/MobileDocumentHeader';
+import DocumentContents from '@components/document/layout/DocumentContents';
+import DocumentFooter from '@components/document/layout/DocumentFooter';
+import DocumentHeader from '@components/document/layout/DocumentHeader';
+import MobileDocumentHeader from '@components/document/layout/MobileDocumentHeader';
 import type {LogParams} from '@type/PageParams.type';
 import markdownToHtml from '@utils/markdownToHtml';
 import {Metadata} from 'next';
