+++ client/src/app/wiki/post/page.tsx
@@ -1,13 +1,13 @@
 'use client';
 
-import PostHeader from '@components/Write/PostHeader';
-import TitleInputField from '@components/Write/TitleInputField';
+import PostHeader from '@components/document/Write/PostHeader';
+import TitleInputField from '@components/document/Write/TitleInputField';
 import {DocumentWriteContextProvider, useDocumentWriteContext} from '@context/DocumentWriteContext';
 import {usePostThrottle} from './usePostThrottle';
-import RelativeSearchTerms from '@components/SearchTerms/RelativeSearchTerms';
+import RelativeSearchTerms from '@components/common/SearchTerms/RelativeSearchTerms';
 
 import {useRelativeSearchTerms} from './useRelativeSearchTerms';
-import TuiEditor from '@components/MarkdownEditor';
+import TuiEditor from '@components/document/MarkdownEditor';
 
 const PostPage = () => {
   const {title, editorRef} = useDocumentWriteContext();
