+++ client/src/components/document/Write/PostHeader.tsx
@@ -1,7 +1,7 @@
 'use client';
 
-import Button from '@components/Button';
-import DocumentTitle from '@components/Document/DocumentTitle';
+import Button from '@components/common/Button';
+import DocumentTitle from '@components/document/layout/DocumentTitle';
 import {useRouter} from 'next/navigation';
 import {useDocumentWriteContext} from '@context/DocumentWriteContext';
