+++ client/tsconfig.json
@@ -18,16 +18,15 @@
         "name": "next"
       }
     ],
-    "baseUrl": "./src",
     "paths": {
-      "@apis/*": ["./apis/*"],
-      "@app/*": ["./app/*"],
-      "@components/*": ["./components/*"],
-      "@constants/*": ["./constants/*"],
-      "@context/*": ["./context/*"],
-      "@hooks/*": ["./hooks/*"],
-      "@type/*": ["./type/*"],
-      "@utils/*": ["./utils/*"]
+      "@apis/*": ["./src/apis/*"],
+      "@app/*": ["./src/app/*"],
+      "@components/*": ["./src/components/*"],
+      "@constants/*": ["./src/constants/*"],
+      "@context/*": ["./src/context/*"],
+      "@hooks/*": ["./src/hooks/*"],
+      "@type/*": ["./src/type/*"],
+      "@utils/*": ["./src/utils/*"]
     }
   },
   "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", ".next/types/**/*.ts"],
